api/edit actions.
