+++ src/vs/workbench/contrib/notebook/browser/contrib/navigation/arrow.ts
@@ -23,6 +23,7 @@ const NOTEBOOK_FOCUS_PREVIOUS_EDITOR = 'notebook.focusPreviousEditor';
 const NOTEBOOK_FOCUS_NEXT_EDITOR = 'notebook.focusNextEditor';
 const FOCUS_IN_OUTPUT_COMMAND_ID = 'notebook.cell.focusInOutput';
 const FOCUS_OUT_OUTPUT_COMMAND_ID = 'notebook.cell.focusOutOutput';
+const CENTER_ACTIVE_CELL = 'notebook.centerActiveCell';
 
 
 registerAction2(class extends NotebookCellAction {
@@ -216,6 +217,27 @@ registerAction2(class extends NotebookCellAction {
 		const editor = context.notebookEditor;
 		const activeCell = context.cell;
 		editor.focusNotebookCell(activeCell, 'editor');
+	}
+});
+
+registerAction2(class CenterActiveCellAction extends NotebookCellAction {
+	constructor() {
+		super({
+			id: CENTER_ACTIVE_CELL,
+			title: localize('notebookActions.centerActiveCell', "Center Active Cell"),
+			keybinding: {
+				when: NOTEBOOK_EDITOR_FOCUSED,
+				primary: KeyMod.CtrlCmd | KeyCode.KEY_L,
+				mac: {
+					primary: KeyMod.WinCtrl | KeyCode.KEY_L,
+				},
+				weight: KeybindingWeight.WorkbenchContrib
+			},
+		});
+	}
+
+	async runWithContext(accessor: ServicesAccessor, context: INotebookCellActionContext): Promise<void> {
+		return context.notebookEditor.revealInCenter(context.cell);
 	}
 });
